requirements.txt
flask
strawberry-graphql
sqlalchemy
databases
gunicorn
